Reorganize code for clarity.
