+++ src/lib/components/app/appSideNav/AppSideNavTree.tsx
@@ -5,13 +5,42 @@ import { VuiIcon } from "../../icon/Icon";
 import { VuiIconButton } from "../../button/IconButton";
 import { VuiAppSideNavLink } from "./AppSideNavLink";
 
+export type Tree = Array<TreeItem>;
+
+export type TreeItem = {
+  name: string;
+  path?: string;
+  pages?: Tree;
+  iconAfter?: React.ReactNode;
+};
+
+export const buildTree = (items: Tree) => {
+  return items.map(({ name, pages, path, iconAfter }) => {
+    if (path) {
+      if (pages) {
+        const childPages = buildTree(pages);
+
+        return (
+          <AppSideNavTreeSection path={path} name={name}>
+            {childPages}
+          </AppSideNavTreeSection>
+        );
+      }
+
+      return <VuiAppSideNavLink path={path} name={name} iconAfter={iconAfter} />;
+    }
+
+    return <div className="vuiAppSideNavTreeSection__info">{name}</div>;
+  });
+};
+
 type Props = {
   name: string;
   path: string;
   children: React.ReactNode;
 };
 
-export const AppSideNavTreeSection = ({ name, path, children }: Props) => {
+const AppSideNavTreeSection = ({ name, path, children }: Props) => {
   const [isOpen, setIsOpen] = useState(true);
 
   const classes = classNames("vuiAppSideNavTreeChildren", {
